Rename modules/helpers/_legacy.py to _constants.py
Truth in labeling: after the 11-PR extraction sprint, this file is 32
lines of pure constants (paths, TTLs, cache dicts, extension sets,
branch overrides) — no functions, no logic. 'Legacy' was accurate when
it was 1613 lines of mixed responsibilities; it isn't now.

Changes:
- git mv _legacy.py -> _constants.py
- Update 15 helpers submodules and 1 test file to import from
  modules.helpers._constants (mechanical sed rewrite)
- Update helpers/__init__.py: 'from ._constants import *' plus a fresher
  comment ('Foundational constants ...')
- Add a module-level docstring to _constants.py listing what it owns
- Update all 25 'extracted from _legacy.py' docstrings across the
  helpers submodules to say 'extracted from the original helpers.py
  monolith' — historically accurate and doesn't reference a filename
  that will keep evolving
- Update two conftest.py comments that mentioned _legacy.py

No behavior change. All 1077 tests pass. No pre-commit issues.
